+++ example/lib/startback_todo.rb
@@ -1,4 +1,8 @@
 require 'startback'
+require 'startback/web/catch_all'
+require 'startback/web/shield'
+require 'startback/web/api'
+require 'startback/web/health_check'
 module StartbackTodo
 end # module StartbackTodo
 require_relative 'startback_todo/errors'