+++ spec/unit/web/test_cors_headers.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'singleton'
+require 'startback/web/cors_headers'
 
 module Startback
   module Web